fix(properties): calculate recursive folder sizes
