+++ src/adapters/directory_summary/tests/bounds.rs
@@ -9,13 +9,17 @@ fn measurement_counts_nested_entries_without_following_symlinks() {
     std::fs::create_dir(root.path().join("nested")).expect("directory");
     std::fs::write(root.path().join("top"), b"abc").expect("top file");
     std::fs::write(root.path().join("nested/child"), b"12345").expect("nested file");
+    std::fs::write(root.path().join("nested/.hidden"), b"hidden").expect("hidden file");
+    std::os::unix::fs::symlink(root.path(), root.path().join("nested/loop")).expect("symlink loop");
+    std::os::unix::fs::symlink(root.path().join("missing"), root.path().join("broken"))
+        .expect("broken symlink");
     std::fs::write(outside.path().join("excluded"), b"excluded content").expect("outside file");
     std::os::unix::fs::symlink(outside.path(), root.path().join("link")).expect("symlink");
     let summary = glib::MainContext::new()
-        .block_on(summarize_trash(&gio::File::for_path(root.path())))
+        .block_on(summarize_directory(&gio::File::for_path(root.path())))
         .expect("summary");
-    assert_eq!(summary.item_count, 4);
-    assert_eq!(summary.total_size, 8);
+    assert_eq!(summary.item_count, 7);
+    assert_eq!(summary.total_size, 14);
     assert!(!summary.truncated);
 }
 
@@ -25,13 +29,13 @@ fn missing_root_is_an_error_and_empty_root_is_exact() {
     let context = glib::MainContext::new();
     assert!(
         context
-            .block_on(summarize_trash(&gio::File::for_path(
+            .block_on(summarize_directory(&gio::File::for_path(
                 root.path().join("missing")
             )))
             .is_err()
     );
     let summary = context
-        .block_on(summarize_trash(&gio::File::for_path(root.path())))
+        .block_on(summarize_directory(&gio::File::for_path(root.path())))
         .expect("empty summary");
     assert_eq!(summary.item_count, 0);
     assert_eq!(summary.total_size, 0);
@@ -51,45 +55,14 @@ fn measurement_budget_is_shared_by_root_files_and_nested_branches() {
     }
     for max_entries in [0, 1, 3, 8] {
         let summary = glib::MainContext::new()
-            .block_on(summarize_trash_with_budget(
+            .block_on(summarize_directory_with_budget(
                 &gio::File::for_path(root.path()),
                 max_entries,
-                MAX_TRASH_DEPTH,
-                TRASH_TIME_BUDGET,
+                MAX_DEPTH,
+                TIME_BUDGET,
             ))
             .expect("bounded summary");
         assert_eq!(summary.item_count, max_entries);
         assert!(summary.truncated);
     }
-}
-
-#[test]
-fn deletion_caps_errors_and_still_processes_successful_siblings() {
-    let root = tempfile::tempdir().expect("fixture");
-    for index in 0..12 {
-        let directory = root.path().join(format!("directory-{index}"));
-        std::fs::create_dir(&directory).expect("directory");
-        std::fs::write(directory.join("keep"), b"not empty").expect("child");
-    }
-    for index in 0..4 {
-        std::fs::write(root.path().join(format!("file-{index}")), b"remove").expect("file");
-    }
-    let mut progress = Vec::new();
-    let outcome = glib::MainContext::new()
-        .block_on(empty_trash(
-            &gio::File::for_path(root.path()),
-            |processed| progress.push(processed),
-        ))
-        .expect("deletion outcome");
-    assert_eq!(outcome.deleted, 4);
-    assert_eq!(outcome.failed, 12);
-    assert_eq!(outcome.errors.len(), 8);
-    assert_eq!(progress.last(), Some(&16));
-    assert!(progress.windows(2).all(|pair| pair[0] < pair[1]));
-    assert_eq!(
-        std::fs::read_dir(root.path())
-            .expect("remaining entries")
-            .count(),
-        12
-    );
 }
